changes underlying harvesting logic
this will now store sources for known rooms to memory, any creep that's set it's source will update the memory to indicate it's harvesting, if it dies deathPanel will remove the creep as a harvester.  Currently the max utilization needs to be manually set on each source's memory.  Started adding a new function to automagically get the max utilization of a source.